feat: add TempDir::new_in/with_prefix/with_prefix_in
Add these functions because they're useful, and to mirror `tempfile`.

These functions can be implemented with older `tempfile`, but it's easier to
bump it up to 3.8 (which is the first version that implemented them). I used
`cargo update --precise` to pin `Cargo.lock` to 3.8.0, but I could just update
`Cargo.lock` and/or `Cargo.toml` to 3.19.1. Let me know!